Text.XML.Stream.Render

Description

Enumeratees to render XML Events. Unlike libxml-enumerator and expat-enumerator, this module does not provide IO and ST variants, since the underlying rendering operations are pure functions.

Documentation

renderBuilder :: Monad m => RenderSettings -> Pipe l Event Builder r m r

Render a stream of Events into a stream of Builders. Builders are from the blaze-builder package, and allow the create of optimally sized ByteStrings with minimal buffer copying.

renderBytes :: MonadUnsafeIO m => RenderSettings -> Pipe l Event ByteString r m r

Render a stream of Events into a stream of ByteStrings. This function wraps around renderBuilder and builderToByteString, so it produces optimally sized ByteStrings with minimal buffer copying.

The output is UTF8 encoded.

renderText :: (MonadThrow m, MonadUnsafeIO m) => RenderSettings -> Pipe l Event Text r m r

Render a stream of Events into a stream of ByteStrings. This function wraps around renderBuilder, builderToByteString and renderBytes, so it produces optimally sized ByteStrings with minimal buffer copying.

data RenderSettings

def :: Default a => a

The default value for this type.

rsPretty :: RenderSettings -> Bool

rsNamespaces :: RenderSettings -> [(Text, Text)]

Defines some top level namespace definitions to be used, in the form of (prefix, namespace). This has absolutely no impact on the meaning of your documents, but can increase readability by moving commonly used namespace declarations to the top level.

rsAttrOrder :: RenderSettings -> Name -> Map Name Text -> [(Name, Text)]

Specify how to turn the unordered attributes used by the Text.XML module into an ordered list.

orderAttrs :: [(Name, [Name])] -> Name -> Map Name Text -> [(Name, Text)]

Convenience function to create an ordering function suitable for use as the value of rsAttrOrder. The ordering function is created from an explicit ordering of the attributes, specified as a list of tuples, as follows: In each tuple, the first component is the Name of an element, and the second component is a list of attributes names. When the given element is rendered, the attributes listed, when present, appear first in the given order, followed by any other attributes in arbitrary order. If an element does not appear, all of its attributes are rendered in arbitrary order.

prettify :: Monad m => Pipe l Event Event r m r

Convert a stream of Events into a prettified one, adding extra whitespace. Note that this can change the meaning of your XML.