Text.XML.Unresolved

Description

DOM-based XML parsing and rendering.

In this module, attribute values and content nodes can contain either raw text or entities. In most cases, these can be fully resolved at parsing. If that is the case for your documents, the Text.XML module provides simplified datatypes that only contain raw text.

Non-streaming functions

writeFile :: RenderSettings -> FilePath -> Document -> IO ()

readFile :: ParseSettings -> FilePath -> IO Document

Lazy bytestrings

renderLBS :: RenderSettings -> Document -> ByteString

parseLBS :: ParseSettings -> ByteString -> Either SomeException Document

parseLBS_ :: ParseSettings -> ByteString -> Document

Text

parseText :: ParseSettings -> Text -> Either SomeException Document

parseText_ :: ParseSettings -> Text -> Document

sinkTextDoc :: MonadThrow m => ParseSettings -> Pipe l Text o u m Document

Byte streams

sinkDoc :: MonadThrow m => ParseSettings -> Pipe l ByteString o u m Document

Streaming functions

toEvents :: Document -> [Event]

fromEvents :: MonadThrow m => Pipe l EventPos o u m Document

renderBuilder :: Monad m => RenderSettings -> Document -> Pipe l i Builder u m ()

renderBytes :: MonadUnsafeIO m => RenderSettings -> Document -> Pipe l i ByteString u m ()

renderText :: (MonadThrow m, MonadUnsafeIO m) => RenderSettings -> Document -> Pipe l i Text u m ()

Exceptions

data InvalidEventStream

Constructors

ContentAfterRoot EventPos

MissingRootElement

InvalidInlineDoctype EventPos

MissingEndElement Name (Maybe EventPos)

UnterminatedInlineDoctype

Settings

def :: Default a => a

The default value for this type.

Parse

data ParseSettings

psDecodeEntities :: ParseSettings -> DecodeEntities

Render

data RenderSettings

rsPretty :: RenderSettings -> Bool

rsNamespaces :: RenderSettings -> [(Text, Text)]

Defines some top level namespace definitions to be used, in the form of (prefix, namespace). This has absolutely no impact on the meaning of your documents, but can increase readability by moving commonly used namespace declarations to the top level.