Text.XML

Description

DOM-based parsing and rendering.

This module requires that all entities be resolved at parsing. If you need to interact with unresolved entities, please use Text.XML.Unresolved. This is the recommended module for most uses cases.

While many of the datatypes in this module are simply re-exported from Data.XML.Types, Document, Node and Element are all redefined here to disallow the possibility of unresolved entities. Conversion functions are provided to switch between the two sets of datatypes.

For simpler, bidirectional traversal of the DOM tree, see the Text.XML.Cursor module.

Data types

data Document

Constructors

Document

Fields

documentPrologue :: Prologue

documentRoot :: Element

documentEpilogue :: [Miscellaneous]

data Prologue

Constructors

Prologue

Fields

prologueBefore :: [Miscellaneous]

prologueDoctype :: Maybe Doctype

prologueAfter :: [Miscellaneous]

data Instruction

Constructors

Instruction

Fields

instructionTarget :: Text

instructionData :: Text

data Miscellaneous

Constructors

MiscInstruction Instruction

MiscComment Text

data Node

Constructors

NodeElement Element

NodeInstruction Instruction

NodeContent Text

NodeComment Text

data Element

Constructors

Element

Fields

elementName :: Name

elementAttributes :: Map Name Text

elementNodes :: [Node]

data Name

A fully qualified name.

Prefixes are not semantically important; they are included only to simplify pass-through parsing. When comparing names with Eq or Ord methods, prefixes are ignored.

The IsString instance supports Clark notation; see http://www.jclark.com/xml/xmlns.htm and http://infohost.nmt.edu/tcc/help/pubs/pylxml/etree-QName.html. Use the OverloadedStrings language extension for very simple Name construction:

myname :: Name
myname = "{http://example.com/ns/my-namespace}my-name"

Constructors

Name

Fields

nameLocalName :: Text

nameNamespace :: Maybe Text

namePrefix :: Maybe Text

data Doctype

Note: due to the incredible complexity of DTDs, this type only supports external subsets. I've tried adding internal subset types, but they quickly gain more code than the rest of this module put together.

It is possible that some future version of this library might support internal subsets, but I am no longer actively working on adding them.

Constructors

Doctype

Fields

doctypeName :: Text

doctypeID :: Maybe ExternalID

data ExternalID

Constructors

SystemID Text

PublicID Text Text

Parsing

Files

readFile :: ParseSettings -> FilePath -> IO Document

Bytes

parseLBS :: ParseSettings -> ByteString -> Either SomeException Document

parseLBS_ :: ParseSettings -> ByteString -> Document

sinkDoc :: MonadThrow m => ParseSettings -> ConduitT ByteString o m Document

Text

parseText :: ParseSettings -> Text -> Either SomeException Document

parseText_ :: ParseSettings -> Text -> Document

sinkTextDoc :: MonadThrow m => ParseSettings -> ConduitT Text o m Document

Other

fromEvents :: MonadThrow m => ConduitT EventPos o m Document

data UnresolvedEntityException

Constructors

UnresolvedEntityException (Set Text)

data XMLException

Constructors

InvalidXMLFile FilePath SomeException

Rendering

writeFile :: RenderSettings -> FilePath -> Document -> IO ()

renderLBS :: RenderSettings -> Document -> ByteString

renderText :: RenderSettings -> Document -> Text

renderBytes :: PrimMonad m => RenderSettings -> Document -> ConduitT i ByteString m ()

Settings

def :: Default a => a

The default value for this type.

Parsing

data ParseSettings

psDecodeEntities :: ParseSettings -> DecodeEntities

psRetainNamespaces :: ParseSettings -> Bool

Whether the original xmlns attributes should be retained in the parsed values. For more information on motivation, see:

https://github.com/snoyberg/xml/issues/38

Default: False

Since 1.2.1

Entity decoding

decodeXmlEntities :: DecodeEntities

Default implementation of DecodeEntities, which leaves the entity as-is. Numeric character references and the five standard entities (lt, gt, amp, quot, pos) are handled internally by the parser.

decodeHtmlEntities :: DecodeEntities

HTML4-compliant entity decoder. Handles the additional 248 entities defined by HTML 4 and XHTML 1.

Note that HTML 5 introduces a drastically larger number of entities, and this code does not recognize most of them.

Rendering

data RenderSettings

rsPretty :: RenderSettings -> Bool

rsNamespaces :: RenderSettings -> [(Text, Text)]

Defines some top level namespace definitions to be used, in the form of (prefix, namespace). This has absolutely no impact on the meaning of your documents, but can increase readability by moving commonly used namespace declarations to the top level.

rsAttrOrder :: RenderSettings -> Name -> Map Name Text -> [(Name, Text)]

Specify how to turn the unordered attributes used by the Text.XML module into an ordered list.

rsUseCDATA :: RenderSettings -> Content -> Bool

Determines if for a given text content the renderer should use a CDATA node.

Default: False

Since: 1.3.3

rsXMLDeclaration :: RenderSettings -> Bool

Determines whether the XML declaration will be output.

Default: True

Since: 1.5.1

orderAttrs :: [(Name, [Name])] -> Name -> Map Name Text -> [(Name, Text)]

Convenience function to create an ordering function suitable for use as the value of rsAttrOrder. The ordering function is created from an explicit ordering of the attributes, specified as a list of tuples, as follows: In each tuple, the first component is the Name of an element, and the second component is a list of attributes names. When the given element is rendered, the attributes listed, when present, appear first in the given order, followed by any other attributes in arbitrary order. If an element does not appear, all of its attributes are rendered in arbitrary order.

Conversion

toXMLDocument :: Document -> Document

fromXMLDocument :: Document -> Either (Set Text) Document

toXMLNode :: Node -> Node

fromXMLNode :: Node -> Either (Set Text) Node

toXMLElement :: Element -> Element

fromXMLElement :: Element -> Either (Set Text) Element